[Datatype] - Add ddl support for FLOAT32 in all the data sources and spanner.

+++ sources/mysql/toddl.go
@@ -85,7 +85,7 @@ func toSpannerTypeInternal(srcType schema.Type, spType string) (ddl.Type, []inte
 		case ddl.String:
 			return ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, []internal.SchemaIssue{internal.Widened}
 		default:
-			return ddl.Type{Name: ddl.Float64}, []internal.SchemaIssue{internal.Widened}
+			return ddl.Type{Name: ddl.Float32}, []internal.SchemaIssue{internal.Widened}
 		}
 	case "numeric", "decimal":
 		switch spType {
